+++ src/providers/namecheap.ts
@@ -1,36 +1,25 @@
 import fetch from 'node-fetch'
-import { getProviderKeys } from '../lib/data'
 import { Provider, ServiceResponse } from '../types/general'
-import { ServiceKey } from '../types/admin'
 import namecheapApi from 'namecheap-api'
 import { providerList } from '.'
+import { getKeys, findKey } from './helpers'
+import _ from 'lodash'
 
 const provider = providerList.find(
   provider => provider.name === 'Namecheap.com'
 )
+
 const { name, dns, keys, service } = provider
 
-const getKeys = (): ServiceKey[] => {
-  const keysDefault: { key: string }[] = [{ key: keys[0] }, { key: keys[1] }]
-  const providerKeys = keysDefault.map(key => {
-    const serviceKeys = getProviderKeys()
-    return serviceKeys.find(k => k.service === dns && k.key === key.key) || key
-  })
-  return providerKeys as ServiceKey[]
-}
-const findKey = (key: string): string => {
-  return (getKeys().find(k => k.key === key) || { value: '' }).value
-}
-
 export const getDomains = async (): Promise<Provider> => {
-  const providerKeys = getKeys()
+  const providerKeys = getKeys(provider)
   const domains = []
 
-  if (findKey(keys[0]) && findKey(keys[1])) {
+  if (findKey(provider, keys[0]) && findKey(provider, keys[1])) {
     const ifconfigRes = await fetch('https://ifconfig.me/ip')
 
-    namecheapApi.config.set('ApiUser', findKey(keys[0]))
-    namecheapApi.config.set('ApiKey', findKey(keys[1]))
+    namecheapApi.config.set('ApiUser', findKey(provider, keys[0]))
+    namecheapApi.config.set('ApiKey', findKey(provider, keys[1]))
     namecheapApi.config.set('ClientIp', await ifconfigRes.text())
 
     const { response } = await namecheapApi.apiCall(
@@ -38,7 +27,7 @@ export const getDomains = async (): Promise<Provider> => {
       {}
     )
 
-    response[0].DomainGetListResult[0].Domain.forEach(domain => {
+    _.get(response, '[0].DomainGetListResult[0].Domain', []).forEach(domain => {
       domains.push({ domain: domain['$'].Name })
     })
   }
